baselines/mpmv2/mltools/requirements.txt
# PyTorch
torch==2.2.1
normflows==1.7.3
torchdiffeq==0.2.3
lightning==2.2.1
flash_attn==2.5.6

# Other deep learning
wandb==0.16.4
onnx==1.15.0
onnxruntime==1.17.1

# Config
omegaconf==2.3.0
hydra-core==1.3.2

# Testing
pytest==8.1.1
coverage==7.4.4

# Standard
matplotlib==3.8.3
scipy==1.12.0
numpy==1.26.4
Pillow==10.2.0
rich==13.7.1
seaborn==0.13.2
tqdm==4.65.0
